Move ISIN job to new service
