+++ backend/src/test/kotlin/blue/mild/covid/vaxx/requests/http-client.env.json
@@ -1,6 +1,8 @@
 {
   "local": {
-    "url": "http://localhost:8080/api"
+    "url": "http://localhost:8080/api",
+    "email": "vaxx@mild.blue",
+    "password": "bluemild"
   },
   "staging": {
     "url": "https://covid-vaxx.stg.mild.blue/api",
